requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jwt==1.3.1
MarkupSafe==2.0.1
pycparser==2.21
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.2
